Fix for https://github.com/mate-desktop/caja/issues/398
On the Shmoocon at 2011 there was a presentation by Jon Larimer demonstrating
how to abuse vulnerabilities and bugs, or even just creating socially or
security compromising thumbnails in mounting and thumbnailing, which happens
on automounting USB drives. This is a particular issue when this happens on a
locked box where the attacker doesn't otherwise have access to the user
account:

http://www.net-security.org/secworld.php?id=10544

Disable automounting if the MATE screen saver is currently locked.

See also https://bugzilla.gnome.org/show_bug.cgi?id=642020

This commit is merely based on :
https://git.gnome.org/browse/nautilus/commit/?h=gnome-2-32&id=b7262fa945ef1ea936c15f0d248ad7a024d97dca