caja-bookmark: don't crash if file is gone
Sometimes the file can be gone without even noticing when
changing the location (and therefore creating the attached bookmark
to that new location).
In bookmark_changed we just disconnected the file if that happened,
so we where failing graciously. But that was not the case when just
changing the location, where we were asserting that the file is not gone
and therefore crashing in such case.

Instead of that, do something similar as bookmark_changed and just
clean up some of the bookmark objects and hope for the best.

origin commit:
https://gitlab.gnome.org/GNOME/nautilus/commit/9429c94