compact view: don't auto-scroll to the right on long filenames
however, now it will auto-scroll to the left if you manually
scroll to the right and then click on any filename. I presume
this behavior is more desired than the previous one. if it's
not so, let me know (in the comments).

adapted from
https://github.com/linuxmint/nemo/commit/72c2214c82637ab622438d46c28e67289a49d472
(thanks to @glebihan)

fixes https://github.com/mate-desktop/caja/issues/406