window-menus: unref extension created action
This looks not very harmful. But definitely is.
Thing is, when using caja_action_from_menu_item it keeps
a reference to the CajaMenuItem from the extension.
So that menu item will never be freed.
Now, let's imagine caja-open-terminal have a ref to the file
that the item points to, and only unref it when the item is destroyed.
Now, sum that when a file is not unrefed completely from caja
when unmounting the file, so it's mark as gone and cannot be used again.
Now try to use it in this state. Caja crashes.

This fix few crashes reported downstream on distros that uses

origin commit:
https://gitlab.gnome.org/GNOME/nautilus/commit/a0cbf72
this extension.