GTK3-stop most desktop flashes in all themes
GTK3-stop desktop from flashing on clicks between desktop and other windows no matter what the theme. This hardcodes the desktop icon view to a transparent background and foreground no matter what the theme.  The black fg color is used in most themes by default and readable in adwaita with dark theme selected. Themes that need to override this can use the #fmdesktopiconview>.view>* selector, I've tested this and it works. Recommend this for BlackMATE, will not bring back flashes. 

Flashes turn black with no other changes made to Caja and now appear only on theme change or heavy disk activity like package installs. If "transparent" is replaced with a color in the background-color setting, the remaining flashes will be that color. Black fg color is sane default, and most themes use it anyway. It seems to appear only as the text color for unselected text when renaming a file on the desktop. A theme can override it without bringing back flashes or failing to override the application priority hardcode by using the #fmdesktopiconview>.view>* selector. Recommend this for BlackMATE.