desktop-canvas-view: scale desktop workarea
The problem is that in the function canvas_container_set_workarea the screen width
and height are in "application pixels" while the workarea ones are in "device
pixels" so when the scaling is > 1, the margins are not properly setted.

We need to scale-down the workarea geometries to "application pixels".

https://bugzilla.gnome.org/show_bug.cgi?id=769302

origin commit:
https://gitlab.gnome.org/GNOME/nautilus/commit/315a55df