GTK3: stop warning spew from sidebar, pathbar
The button underallocation warnings came from pathbar sliders picking up padding from themes, then being forced to render to a smaller size in the code.

The GTK_ICON_LOOKUP_GENERIC_FALLBACK warnings came from the sidebar icons, this flag is now a source of errors. Unfortunately NULL does not work, so set GTK_ICON_LOOKUP_FORCE_SIZE, which does not create errors and ensures the icon size we actually want in the sidebar is used anyway.

Stops all but two of the warnings listed in
https://github.com/mate-desktop/caja/issues/561