CSS: Load a different caja-desktop.css with GTK 3.14
In caja-desktop.css this line isn't compatible with GTK 3.14:

".caja-desktop:not(:selected):not(:active):not(.rubberband){"

It is needed for GTK 3.20/3.22 tough.

This PR checks the GTK version and loads the appropriate CSS.

This fixes desktop icons showing black text. That issue was
reproduced with Adwaita under LMDE with GTK 3.14.