private: Ensure we allow launching .desktop files from trusted symlinks
When a file is a symlink to one of the XDG data dirs, we'll allow that
link to work, as they're vendor provided and not world-writeable by a
malicious entity.

This lookup is handled by the "is_system_dir" logic, to ensure that whatever
the .desktop symlink is pointing at is already implicitly whitelisted. This
allows for vendor provided "default" symlinks on within user directories,
such as installer shortcuts for LiveCDs, etc.

Signed-off-by: Ikey Doherty <ikey@solus-project.com>