Add sort criterion by reversed extension segments
The basenames of files are split by dots and then starting from the end
each segment is compared to find a sort order. Example:

    bar.tar.bz2
    foo.tar.bz2
    a.bar.gz
    x.tar.gz
    z.tar.gz
    test.tex

A heuristic determines what extension segments are part of the
extension. There is probably no perfect solution but there are much
less false positives. As a result the sorting is more intuitive and the
displayed column in the list view is better readable and displays
extensions more accurately.

In addition a bug related to the default sort criteria in the
preferences has been fixed.