Eliminate accessible children-changed event flood during update all
Accessible children-changed events should only be emitted when a user-
triggered action (creation of file, copying an item) causes an item to
be added.

Set container->details->is_loading (which does not seem to be used
anywhere) in caja_icon_container_request_update_all because if the
entire container is being updated, children-changed events should not
be emitted for each updated icon.