desktop window: Add is-desktop AtkObject attribute
Screen readers need to distinguish between the desktop window and normal
windows, to be able to provide nicer speech synthesis.

This can be done by simply adding an "is-desktop" attribute to the
underlying AtkObject.

This is here done by introducing a thin caja_desktop_window_accessible_class
class which is based on GtkWindowAccessible and just appends the
attribute.

Closes: #999