[libcaja-private] Re-check file MIME type before picking an application.
An example of when the MIME type might change: a file is initially created
with 0 bytes of content, but more data is added later. Empty files are always
detected as plain text, but the file might not be empty anymore when the user
opens it.

This commit affects the behavior when double-clicking a file and when right-
clicking on it too.