desktop: Always draw the items label frame
Draw the frame under the items label even when not selected and not in
prelight state, so that it is possible to theme it in the other states.
However we still don't draw it when editing, because there is the
editing widget frame already and drawing the label's one might slightly
overflow the editing widget's borders and lead to slightly ugly
results.

Being able to draw a background is especially useful for accessibility
purposes (e.g. a high contrast theme would typically make sure there is
a background with enough contrast), but might be useful to all as
desktop backgrounds are often not uniform at all.

By default this is drawn transparent on the non-selected and
non-prelight states (no change in default appearance then), but allows
the user to customize it with a background if need be.

Unfortunately for the moment this cannot be altered by theme, so the
user will have to modify the GTK custom user CSS directly, but it still
is better than nothing and can hopefully be solved better later.