Restore transparent background on icon view labels
18e4fbd005bd561e4a12ed388d4d9a8c93f0b15b started always drawing the
frame on canvas items for the desktop's sake, but it also impacts other
icon views.  This is good in theory, but conflicts with some themes
that used to work fine and now display an unwanted background because
of generic rules matching these elements.

Unfortunately there don't seem to be another solution than providing
overrides for this yet, so do that, as anyway before recently the
background wasn't drawn at all even if specified.