Show non-generic mime type icon where applicable
When the target mime type icon is in 'hicolor', themed generic icon is
currently preferred.

The reason is: the gThemedIcon for the file mentions the generic icon
name as the last alternative. Themed icon name selection prefers it over
the others because it is always in the theme, while others are not.

The solution implemented here is to reduce the gThemedIcon names to the
first usable one only, in hope the generic icon name will be dropped. If
no usable icon is found, the original gThemedIcon is kept unchanged.

Fixes #795