Navigation windows:  Save the state of the extra pane
Already, the presence and position of the sidebar in navigation
windows is recorded, and all future navigation windows will open
up by default with the user's chosen sidebar preferences.  This
commit gives that same flexibility to the extra pane:  If the
user likes the extra pane, when they open it, Caja will record
to open the extra pane by default; and it will also record the
position of the handle of the extra pane, so that if the user
likes the primary pane to take up 40% of the navigation area
and the extra pane 60%, then Caja will open up the panes to that
position by default.

Requested by @tatanka on the Ubuntu MATE Community:
https://ubuntu-mate.community/t/feature-request-caja/24937