Use a fraction to express position of extra pane
Previously, the position of a Caja navigation window's extra pane
was expressed as an offset measured in pixels.  The pixel-based
offset does not necessarily scale well to different window or
monitor sizes, so this commit changes the measurement to a fraction
between 0 and 1.  0.5 means that the two panes of the window are
equally sized.