caja-file.c: fix crash with empty server name
When an SMB server has an empty name, details->display_name is NULL.
This causes a couple of different issues that both result in crashes:

1. When update_info_internal fills in details->name, it inadvertently
uses NULL instead of "" because eel_strcmp treats "" and NULL as
identical. This results in an assertion failure later on in
caja_directory_add_file. The code now ensures that details->name won't
be filled in as NULL. This also allows us to switch to normal strcmp.

2. After the above issue is fixed, there's still a crash due to a bug
that was also already fixed in Nautilus. This fix ensures that
caja_file_peek_display_name doesn't return NULL. See:
https://gitlab.gnome.org/GNOME/nautilus/-/commit/3fb7cff1a

Fixes #689