Make CajaIconData type less confusing
This is actually an opaque type, and there is no definition of struct
CajaIconData anywhere.  This actually doesn't change anything as the
non-existent struct is equivalent to any other incomplete type, and
works fine so long as the pointers are not dereferenced using that
incomplete type.

However, change this to an explicit `void` to make it clear it's an
opaque pointer and stop people from looking for a struct CajaIconData
that is nowhere to be found.