Enable showing in Budgie Desktop and XFCE
This PR enables the displaying of Caja (as a file manager) for use in Budgie Desktop and XFCE. Prior to this commit, it would only be able accessible in MATE or with downstreams like Fedora patching in the support. The addition of these is important as file managers such as Nautilus adopt libadwaita and begin to directly come into conflict with user theming choices.