caja-file-operations: restart timer also for moves
As done for copy/delete jobs, restart the operation timer also for move
jobs. Thus preparations (e.g. `scan_sources`) later won't affect the
later transfer rate calculation.

caja-file-operations: stop timer when waiting

When a copy/move operation is created while another operation is already
active, the new operation is queued. As the (already running) operation
timer of the new operation is not stopped during the waiting period,
that period is (erroneously) included in the transfer rate calculation
and leads to initially low/slowly increasing transfer rates be shown.

Hence stop the operation timer when the (queued) operation is waiting.

Fixes #1420 and #1623.