eel-labeled-image: Implement a11y for button types
Add accessible implementation for EelLabeledImageButton,
EelLabeledImageCheckButton, EelLabeledImageToggleButton and
EelLabeledImageRadioButton.

The implementation is actually the same as the one for EelLabeledImage
which already handles the button cases, this just plugs in the right
bits for those other types to use it properly.

This fixes accessibility of the emblem selector in the file properties
window as well as the groups in the "property borwser" (the
"Backgrounds & Emblems" window).