File properties: improve accessibility of the simple permissions
As the UI doesn't have semantic groups for owner/group/others settings
but they are only visually apparent, add some more context to the
accessible version of the labels to identify what it targets.

This is especially useful for the setting for other's permissions,
because the focus never lands on the line with the "Others" label,
hence making it hard to infer the context while navigating.

Another solution would be to redesign the UI to have semantic groups,
like e.g. labeled frames, but it's a lot less trivial and a lot more
intrusive a change.