WIP tech-preview git branch icon text - NOT FOR MERGE
Display the git branch along with directory name in icon view; when the directory is a git repository.

Technology preview only in order to gauge appetite for mainline integration.

As git is becoming increasingly ubiquitous such as feature may be deemed appropriate for internal adoption. Performance impact should be negligible.

No current extension found which is fit for purpose, and it was deemed easier and more logical to propose direct integration.

Obviously further work will be required to enable/disable as a displayed option, and potentially some level of customisation. This would require open discussion of possibilities.

Solutions for co-existence with other displayable parameters also to be community determined.

Consideration and feedback appreciated.