New feature - show git branch with icon view text (#1838)
Display the git branch along with directory name in icon view; when the directory is a git repository.

Obviously further work will be required to enable/disable as a displayed option, and potentially some level of customisation. This would require open discussion of possibilities.

Solutions for co-existence with other displayable parameters also to be community determined.

Consideration and feedback appreciated.

* Add UI support to enable/disable option

Added user interface support for option to enable and disable the display of the git branch name in icon view.

* Add internationsation support for the new feature

With translations for most of the currently supported languages.

* Update caja.pot

Build system regenerated .pot file with additional strings from new feature (already in previous commit), new strings already introduced to code-base, along with updated source file line numbers of existing strings. Required for the automated pull mechanism in Transifex - which should complete once merged with master branch.